+++ lib/system_test/src/code_completion_result.cpp
@@ -114,5 +114,10 @@ namespace metashell
         return false;
       }
     }
+
+    bool operator!=(const code_completion_result& r_, const json_string& s_)
+    {
+      return !(r_ == s_);
+    }
   }
 }